requirements.txt
django==3.2
djangorestframework
# mollie-api-python # for dummpy payment
django-cors-headers
gunicorn